Change benchmark to run on linux
A lot of small temporary changes to make the benchmark run. Changed the schedulers best plan chooser to use more accurate configuration and execution times.

+++ src/dbmstodspi/query_execution/scheduling/scheduling_data.hpp
@@ -33,7 +33,7 @@ namespace orkhestrafs::dbmstodspi::scheduling_data {
 struct ExecutionPlanSchedulingData {
   std::unordered_set<std::string> processed_nodes;
   std::map<std::string, TableMetadata> data_tables;
-  int streamed_data_size;
+  long streamed_data_size;
 };
 
 }  // namespace orkhestrafs::dbmstodspi::scheduling_data
